+++ CHANGELOG.md
@@ -5,6 +5,8 @@ ENHANCEMENTS:
 
 * resource/`junos_security_dynamic_address_feed_server`: add `url`, `tls_profile` and `validate_certificate_attributes_subject_or_san` arguments. (Fixes #381)  
   Now one of `hostname` or `url` arguments is required.
+* resource/`junos_interface_logical`: add `tunnel` block argument (Fixes #383)
+* data-source/`junos_interface_logical`: add `tunnel` block attribute (like resource)
 
 BUG FIXES: